Co-creation-projects/Tangent-90C-ChatTest/pyproject.toml
[project]
name = "camel-exam-agent"
version = "0.1.0"
description = "一个基于Camel-AI的考研复试出题Agent，能够帮助用户高效准备考研复试。"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "camel-ai[data-tools,document-tools,rag,storage]>=0.2.82",
    "httpx[socks]>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
]
